backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
anthropic==0.49.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==23.0.0
google-genai>=1.0.0
PyJWT>=2.8.0
requests>=2.31.0
redis>=5.0.0
ffmpeg-python>=0.2.0
imageio-ffmpeg>=0.5.1
pdfplumber>=0.11.0
Pillow>=10.0.0
pypdfium2>=4.30.0
